# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-refund-hold",
    "version": "0.1.0",
    "description": "The refundable-hold pattern for x402 merchants, as drop-in Express middleware. Charge a hold, auto-refund on failure, ledger included. Pay in USDC on Base or Solana — the client picks the rail.",
    "license": { "name": "Apache-2.0", "identifier": "Apache-2.0" },
    "contact": { "name": "nirholas", "email": "nichxbt@gmail.com", "url": "https://github.com/nirholas/x402-refund-hold" }
  },
  "servers": [
    { "url": "http://localhost:4021", "description": "Local dev server" }
  ],
  "tags": [
    { "name": "paid", "description": "Routes behind the x402 dual-rail paywall" },
    { "name": "free", "description": "Free routes" }
  ],
  "paths": {
    "/demo/book": {
      "post": {
        "tags": ["paid"],
        "operationId": "demoBook",
        "summary": "Book a demo reservation behind a $0.01 refundable hold",
        "description": "Price: $0.01 USDC. On success returns the confirmation plus the signed captured hold. On any failure the payment is refunded and a signed refund record is returned in the same 200 body.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BookingRequest" },
              "examples": {
                "success": { "value": { "name": "Agent Guest", "partySize": 2 } },
                "failure": { "value": { "simulate": "failure" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Confirmed booking, or a refund record when the booking failed.",
            "headers": {
              "X-PAYMENT-RESPONSE": { "$ref": "#/components/headers/XPaymentResponse" },
              "X-Hold-Id": { "schema": { "type": "string" }, "description": "Ledger id of the hold opened for this request" }
            },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BookingOutcome" } }
            }
          },
          "402": { "$ref": "#/components/responses/PaymentRequired" },
          "500": { "$ref": "#/components/responses/ServerError" },
          "502": { "$ref": "#/components/responses/FacilitatorError" }
        }
      }
    },
    "/holds/{id}": {
      "get": {
        "tags": ["free"],
        "operationId": "getHold",
        "summary": "Hold status from the refund ledger",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "hold_9f0c2c0e-1f2a-4c1e-9c0e-1f2a4c1e9c0e" }
        ],
        "responses": {
          "200": {
            "description": "The hold",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "hold": { "$ref": "#/components/schemas/Hold" } }, "required": ["hold"] }
              }
            }
          },
          "404": {
            "description": "Unknown hold id",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/verify": {
      "post": {
        "tags": ["free"],
        "operationId": "verifyRecord",
        "summary": "Verify the HMAC signature on a record issued by this service",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["payload", "signature"],
                "properties": { "payload": {}, "signature": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": { "type": "object", "required": ["valid"], "properties": { "valid": { "type": "boolean" } } }
              }
            }
          },
          "400": {
            "description": "Missing payload or signature",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": ["free"],
        "operationId": "health",
        "summary": "Liveness and the payment rails currently advertised",
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "service": { "type": "string" },
                    "rails": { "type": "array", "items": { "$ref": "#/components/schemas/Rail" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/.well-known/x402": {
      "get": {
        "tags": ["free"],
        "operationId": "manifest",
        "summary": "x402 discovery manifest",
        "responses": {
          "200": {
            "description": "Manifest describing paid resources, prices and both payment rails",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    }
  },
  "components": {
    "headers": {
      "XPaymentResponse": {
        "description": "Base64 JSON settlement receipt: { success, rail, network, transaction, payer, amount, asset, resource }",
        "schema": { "type": "string" }
      }
    },
    "responses": {
      "PaymentRequired": {
        "description": "Payment required. The `accepts` array lists both rails — pick one, sign it, and retry with an `X-PAYMENT` header.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/PaymentRequiredBody" },
            "example": {
              "x402Version": 1,
              "error": "X-PAYMENT header is required",
              "accepts": [
                {
                  "scheme": "exact",
                  "network": "base-sepolia",
                  "maxAmountRequired": "10000",
                  "resource": "http://localhost:4021/demo/book",
                  "description": "x402-refund-hold: POST /demo/book",
                  "mimeType": "application/json",
                  "payTo": "0x40252CFDF8B20Ed757D61ff157719F33Ec332402",
                  "maxTimeoutSeconds": 60,
                  "asset": "0x036CbD53842c5426634e7929541eC2318f3dCF7e",
                  "extra": { "name": "USDC", "version": "2" }
                },
                {
                  "scheme": "exact",
                  "network": "solana",
                  "maxAmountRequired": "10000",
                  "resource": "http://localhost:4021/demo/book",
                  "description": "x402-refund-hold: POST /demo/book",
                  "mimeType": "application/json",
                  "payTo": "WwwuGbqHrwF5RG89KhUbmRWEvjnRH9k5kVM5p7T3WwW",
                  "maxTimeoutSeconds": 60,
                  "asset": "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v",
                  "extra": { "name": "USD Coin", "decimals": 6, "feePayer": "2wKupLR9q6wXYppw8Gr2NvWxKBUqm4PPJKkQfoxHDBg4", "amount": "10000" }
                }
              ]
            }
          }
        }
      },
      "ServerError": {
        "description": "Server misconfiguration (e.g. no valid payTo on either rail)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "FacilitatorError": {
        "description": "The payment facilitator could not be reached. Nothing was charged.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "PaymentRequirements": {
        "type": "object",
        "required": ["scheme", "network", "maxAmountRequired", "resource", "payTo", "asset"],
        "properties": {
          "scheme": { "type": "string", "const": "exact" },
          "network": { "type": "string", "examples": ["base-sepolia", "base", "solana", "solana-devnet"] },
          "maxAmountRequired": { "type": "string", "description": "USDC atomic units (6 decimals)" },
          "resource": { "type": "string", "format": "uri" },
          "description": { "type": "string" },
          "mimeType": { "type": "string" },
          "payTo": { "type": "string" },
          "maxTimeoutSeconds": { "type": "integer" },
          "asset": { "type": "string", "description": "USDC contract address (EVM) or SPL mint (Solana)" },
          "extra": { "type": "object", "additionalProperties": true }
        }
      },
      "PaymentRequiredBody": {
        "type": "object",
        "required": ["x402Version", "accepts"],
        "properties": {
          "x402Version": { "type": "integer", "const": 1 },
          "error": { "type": "string" },
          "accepts": { "type": "array", "items": { "$ref": "#/components/schemas/PaymentRequirements" }, "minItems": 1 }
        }
      },
      "Rail": {
        "type": "object",
        "properties": {
          "rail": { "type": "string", "enum": ["evm", "solana"] },
          "network": { "type": "string" },
          "payTo": { "type": "string" },
          "facilitator": { "type": "string", "format": "uri" }
        }
      },
      "BookingRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "time": { "type": "string", "format": "date-time" },
          "partySize": { "type": "number" },
          "simulate": { "type": "string", "enum": ["success", "failure", "reject"], "default": "success" }
        }
      },
      "BookingConfirmation": {
        "type": "object",
        "properties": {
          "reservationId": { "type": "string" },
          "name": { "type": "string" },
          "confirmedTime": { "type": "string", "format": "date-time" },
          "partySize": { "type": "number" },
          "venue": { "type": "string" },
          "bookedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Hold": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "resource": { "type": "string" },
          "payer": { "type": "string" },
          "amountUsd": { "type": "number" },
          "network": { "type": "string" },
          "rail": { "type": "string", "enum": ["evm", "solana", "none"] },
          "settled": { "type": "boolean" },
          "paymentTx": { "type": "string" },
          "status": { "type": "string", "enum": ["held", "captured", "voided", "refunded", "settled"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "refundableUntil": { "type": "string", "format": "date-time" },
          "reason": { "type": "string" },
          "refundTxHash": { "type": "string" },
          "refundMode": { "type": "string", "enum": ["onchain", "claim"] },
          "artifact": {}
        }
      },
      "RefundRecord": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "const": "refund" },
          "holdId": { "type": "string" },
          "payer": { "type": "string" },
          "amountUsd": { "type": "number" },
          "network": { "type": "string" },
          "rail": { "type": "string", "enum": ["evm", "solana", "none"] },
          "reason": { "type": "string" },
          "mode": { "type": "string", "enum": ["onchain", "claim"] },
          "txHash": { "type": "string" },
          "issuedAt": { "type": "string", "format": "date-time" }
        }
      },
      "SignedHold": {
        "type": "object",
        "properties": {
          "payload": { "$ref": "#/components/schemas/Hold" },
          "signature": { "type": "string" },
          "algorithm": { "type": "string", "const": "HMAC-SHA256" }
        }
      },
      "SignedRefund": {
        "type": "object",
        "properties": {
          "payload": { "$ref": "#/components/schemas/RefundRecord" },
          "signature": { "type": "string" },
          "algorithm": { "type": "string", "const": "HMAC-SHA256" }
        }
      },
      "BookingOutcome": {
        "type": "object",
        "required": ["outcome"],
        "properties": {
          "outcome": { "type": "string", "enum": ["confirmed", "refunded", "rejected"] },
          "confirmation": { "$ref": "#/components/schemas/BookingConfirmation" },
          "hold": { "$ref": "#/components/schemas/SignedHold" },
          "refund": { "$ref": "#/components/schemas/SignedRefund" },
          "reason": { "type": "string" },
          "refundTerms": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "hint": { "type": "string" },
          "holdId": { "type": "string" }
        }
      }
    }
  }
}